cfg unit test refactoring
