+++ tests/conftest.py
@@ -46,7 +46,7 @@ use_grpc = os.getenv("USE_GRPC") in {"1", True}
 
 config = {
     "desktopVersion": "2026.1",
-    "use_grpc": use_grpc,
+    "use_grpc": True,
 }
 
 # Check for the local config file, override defaults if found
